stream-ingest/requirements.txt
google-cloud-pubsub==1.2.0
Flask==1.1.1
pytest==5.4.3
slackeventsapi
